requirements.txt
tkinterdnd2>=0.3.0
pillow